requirements.txt
pytest>=8
PyYAML>=6
pygame>=2.6
